Revert "[Add new Entities][Add Services: -Update Status, Priority, and TaskType]"
