+++ src/todos/dto/todo.ts
@@ -20,17 +20,43 @@ export class CreateTodoDto {
 
   @IsOptional()
   @IsEnum(Priority)
-  priority?: Priority = Priority.LOW;
+  priority?: Priority = Priority.MEDIUM;
 
   @IsOptional()
   @IsEnum(Status)
   status?: Status = Status.TODO;
 }
 
+export class UpdateTodoDto {
+  @IsOptional()
+  @IsString()
+  @MaxLength(255)
+  title?: string;
 
+  @IsOptional()
+  @IsString()
+  description?: string;
+
+  @IsOptional()
+  @IsEnum(Priority)
+  priority?: Priority;
+
+  @IsOptional()
+  @IsEnum(Status)
+  status?: Status;
+}
 
 export class DeleteTodoDto {
   @IsNumber()
   id: number;
 }
 
+export class TodoDto {
+  id: number;
+  title: string;
+  description?: string;
+  priority: Priority;
+  status: Status;
+  createdAt: Date;
+  updatedAt: Date;
+}
