--- src/todos/enums/taskType.ts
@@ -1,4 +1,0 @@
-export enum TaskType {
-  Scheduled = "scheduled",
-  Unscheduled = "unscheduled",
-}
