+++ src/todos/todos.entity.ts
@@ -1,5 +1,5 @@
 // todos/todos.entity.ts
-import { IsBoolean, IsEnum, IsOptional, IsString } from "class-validator";
+import { IsEnum, IsOptional, IsString } from "class-validator";
 import {
   Column,
   CreateDateColumn,
@@ -40,26 +40,9 @@ export class Todo {
   @IsEnum(Status)
   status: Status;
 
-  @Column({
-    type: "boolean",
-    default: true,
-  })
-  isScheduled: boolean;
-
   @CreateDateColumn()
   createdAt: Date;
 
   @UpdateDateColumn()
   updatedAt: Date;
-
-  @Column({ type: "timestamp", nullable: true })
-  completedAt: Date;
-  setStatus(status: Status) {
-    if (status === "completed") {
-      this.completedAt = new Date();
-    } else {
-      this.completedAt = null;
-    }
-  }
 }
-
